backend/requirements.txt
# MentorMind Backend Dependencies
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
aiohttp>=3.9.0

# Auth
PyJWT>=2.8.0
cryptography>=41.0.0
requests>=2.31.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0  # For PostgreSQL
alembic>=1.12.0  # For database migrations

# AI/ML
numpy>=1.24.0
networkx>=3.0
opencv-python>=4.8.0
manim>=0.18.0  # Math animation engine for video generation

Pillow>=10.0.0
scikit-image>=0.22.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Cloud & Queues
boto3==1.34.0
celery==5.3.6
redis==5.0.1
python-multipart>=0.0.6

# Speech Recognition
funasr>=1.1.0
modelscope>=1.9.0
openai-whisper>=20231117
# OCR
paddleocr>=2.7.0
paddlepaddle>=2.5.0
opencv-python-headless>=4.8.0